Update from-gatsby.md
The original instructions when converting from gatsby to next had users wrap the inner HTML inside the "Link" in an "a" tag, this is incorrect as of version 13+ and throws an error. I just refactored my project as per the instructions and am now going through to revert it. Tweaked wording and example for the correct and up dated process.